backend_blog/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1